changed PageID to int64, and pageSize to int8

+++ requests/accounts_requests.go
@@ -10,8 +10,8 @@ type GetAccountRequest struct {
 }
 
 type GetAccountsListRequest struct {
-	PageID   uint64 `form:"page_id" binding:"required,min=1"`
-	PageSize uint64 `form:"page_size" binding:"required,min=5,max=10"`
+	PageID   int64 `form:"page_id" binding:"required,min=1"`
+	PageSize int8  `form:"page_size" binding:"required,min=5,max=10"`
 }
 
 type DepositRequest struct {
